check if token exists

+++ simple-node-server/routes/api/users.js
@@ -96,12 +96,16 @@ router.post('/login', async (req, res) => {
 
 // Delete user
 router.delete('/:username', async (req, res) => {
-  try {
-    await User.deleteOne({ username: req.params.username });
-    await Profile.deleteOne({ username: req.params.username });
-    res.send({ status: 'ok' });
-  } catch (e) {
-    res.send({ error: e });
+  if (req.body.token.length > 0) {
+    try {
+      await User.deleteOne({ username: req.params.username });
+      await Profile.deleteOne({ username: req.params.username });
+      res.send({ status: 'ok' });
+    } catch (e) {
+      res.send({ error: e });
+    }
+  } else {
+    res.send({ error: 'Error: token does not exist' });
   }
 });
